requirements.txt
streamlit==0.61.0
pandas==1.1.3
plotly==4.11.0
requests==2.11.1
Pillow
pip install opencv-python

